+++ src/Traits/IsPresentable.php
@@ -4,19 +4,24 @@ declare(strict_types=1);
 
 namespace TPG\IsPresentable\Traits;
 
-use Illuminate\Support\Collection;
-use Illuminate\Support\Str;
-use ReflectionClass;
-use ReflectionMethod;
+use TPG\IsPresentable\IsPresentableService;
 use TPG\IsPresentable\Presenter;
 
 trait IsPresentable
 {
-    protected $presentables = [];
-
     public function presentable(): Presenter
     {
-        return new Presenter($this->getPresentables());
+        return new Presenter($this->instance()->getPresentables($this));
+    }
+
+    public function setPresentables(array $presentable): void
+    {
+        $this->presentables = $presentable;
+    }
+
+    public function getPresentables(): array
+    {
+        return $this?->presentables ?? [];
     }
 
     public function toArray(): array
@@ -24,7 +29,7 @@ trait IsPresentable
         return array_merge(
             $this->getOriginalAttributes(),
             [
-                config('presentable.key') => $this->getPresentables(),
+                config('presentable.key') => $this->instance()->getPresentables($this, true),
             ],
         );
     }
@@ -38,23 +43,8 @@ trait IsPresentable
         return [];
     }
 
-    protected function getPresentables(): array
+    protected function instance(): IsPresentableService
     {
-        return $this->getPresentableMethods()->mapWithKeys(function (ReflectionMethod $method) {
-            $name = Str::after($method->name, 'presentable');
-
-            return [Str::snake($name) => $this->{'presentable'.$name}()];
-        })->toArray();
-    }
-
-    protected function getPresentableMethods(): Collection
-    {
-        $reflection = new ReflectionClass($this);
-
-        return collect($reflection->getMethods())
-            ->filter(
-                fn (ReflectionMethod $method) => $method->name !== 'presentable'
-                    && Str::startsWith($method->name, 'presentable')
-            );
+        return app('is-presentable');
     }
 }