+++ tests/IsPresentableTest.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace TPG\Tests;
 
+use Illuminate\Database\Eloquent\Model;
+use TPG\IsPresentable\Traits\IsPresentable;
+
 class IsPresentableTest extends TestCase
 {
     /**
@@ -30,9 +33,60 @@ class IsPresentableTest extends TestCase
         self::assertSame([
             'name' => 'Test User',
             'presentable' => [
+                'created_at' => now()->format('d F Y H:i a'),
                 'test' => 'presentable-test',
             ],
         ], $user->toArray());
+    }
+
+    /**
+     * @test
+     **/
+    public function class_presenters_can_have_options(): void
+    {
+        $user = new class extends Model
+        {
+            use IsPresentable;
+
+            protected $guarded = [];
+
+            protected array $presentables = [
+                'options' => [OptionPresenter::class, ['option 1', 'option 2']],
+            ];
+        };
+
+        $this->assertSame('option 1 + option 2', $user->presentable()->options);
+    }
+
+    /**
+     * @test
+     **/
+    public function class_presenters_can_use_the_attribute(): void
+    {
+        $user = new class extends Model
+        {
+            use IsPresentable;
+
+            protected array $presentables = [
+                'name' => AttributePresenter::class,
+                'age' => AttributePresenter::class,
+            ];
+
+            public function getNameAttribute(): string
+            {
+                return 'slim';
+            }
+
+            public function getAgeAttribute(): string
+            {
+                return 'sixty';
+            }
+        };
+
+        $this->assertSame([
+            'name' => 'slim',
+            'age' => 'sixty',
+        ], $user->toArray()['presentable']);
     }
 
     /**
@@ -49,6 +103,7 @@ class IsPresentableTest extends TestCase
         self::assertSame([
             'name' => 'Test User',
             'test-key' => [
+                'created_at' => now()->format('d F Y H:i a'),
                 'test' => 'presentable-test',
             ],
         ], $user->toArray());
@@ -64,5 +119,21 @@ class IsPresentableTest extends TestCase
         ]);
 
         self::assertNull($user->presentable()->nothing);
+    }
+
+    /**
+     * @test
+     **/
+    public function presentables_can_be_cast_as_an_array(): void
+    {
+        $user = new User([
+            'name' => 'Test User'
+        ]);
+
+        self::assertSame([
+            'created_at' => now()->format('d F Y H:i a'),
+            'hidden' => 'hidden',
+            'test' => 'presentable-test',
+        ], $user->presentable()->toArray());
     }
 }