+++ tests/User.php
@@ -13,6 +13,11 @@ class User extends Model
 
     protected $guarded = [];
 
+    protected array $presentables = [
+        'created_at' => CreatedAtPresenter::class,
+        'hidden' => HiddenPresenter::class,
+    ];
+
     public function presentableTest(): string
     {
         return 'presentable-test';