+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] 02-01-2022
+### New year, new release!
+- Version 2.0.0 released.
+
+## [2.0.0-beta.2] 23-12-2021
+### Additions
+- The `Presenter` class now has a `toArray()` method which will return ALL presentables as an array (including hidden presentables).
+
+## [2.0.0-beta.1] 14-12-2021
+### Additions
+- Model classes can now specify presentable classes by passing the attribute name and the class path to the `$presentables` array attribute.
+- If the class cannot be found, a new `InvalidPresentableClass` exception will be thrown.
+- Added the ability to hide presenters from the array by implementing the `IsHidden` interface.
+- Added the presenter options. Options can be passed to presenter classes by passing an array to each item in the `$presentables` array.
+- Added the ability for presenter classes to have access to the attribute from the `$presentables` array.
+- Default presentables can be added to the `presentable` config file and will be added to all `IsPresentable` classes.
+
 ## [1.1.0] 06-07-2021
 ### Changes
 - The signature for the `Presenter` class has been changed. The first `Model` parameter has been removed as it was never actually used.